requirements.txt
appnope==0.1.3
asttokens==2.2.1
backcall==0.2.0
colorama==0.4.6
comm==0.1.3
debugpy==1.6.7
decorator==5.1.1
executing==1.2.0
importlib-metadata==6.6.0
ipykernel==6.23.1
ipython==8.13.2
jedi==0.18.2
jupyter_client==8.2.0
jupyter_core==5.3.0
matplotlib-inline==0.1.6
nest-asyncio==1.5.6
packaging==23.1
parso==0.8.3
pexpect==4.8.0
pickleshare==0.7.5
platformdirs==3.5.1
prompt-toolkit==3.0.38
psutil==5.9.5
ptyprocess==0.7.0
pure-eval==0.2.2
pyasn1==0.5.0
pycryptodome==3.18.0
pycryptodomex==3.18.0
Pygments==2.15.1
python-dateutil==2.8.2
pyzmq==25.1.0
rsa==4.9
six==1.16.0
stack-data==0.6.2
tornado==6.3.2
traitlets==5.9.0
typing_extensions==4.6.2
wcwidth==0.2.6
zipp==3.15.0
